Update pow worker code
